% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-methods.R
\name{print.ard_glad}
\alias{print.ard_glad}
\alias{plot.ard_glad}
\alias{as.list.ard_glad}
\alias{ard_glad_sources}
\alias{ard_glad_sources.default}
\alias{ard_glad_sources.ard_glad}
\alias{new_source_names}
\alias{new_source_names.default}
\alias{new_source_names.ard_glad}
\title{print method for ard_glad objects}
\usage{
\method{print}{ard_glad}(x, ...)

\method{plot}{ard_glad}(
  x,
  r = 3,
  g = 2,
  b = 1,
  zlim = c(0, 5000),
  stretch = "lin",
  colNA = "#acacac",
  ...
)

\method{as.list}{ard_glad}(x, ...)

ard_glad_sources(x, ...)

\method{ard_glad_sources}{default}(x, ...)

\method{ard_glad_sources}{ard_glad}(x, ...)

new_source_names(x, ...)

\method{new_source_names}{default}(x, ...)

\method{new_source_names}{ard_glad}(x, label, ...)
}
\arguments{
\item{x}{an ard_glad object}

\item{...}{not used}

\item{r}{integer between 1 and \code{nlyr(x)}. Layer to use as the Red channel}

\item{g}{integer between 1 and \code{nlyr(x)}. Layer to use as the Green channel}

\item{b}{integer between 1 and \code{nlyr(x)}. Layer to use as the Blue channel}

\item{zlim}{numeric vector of length 2. Range of values to plot (optional). If this is set, and \code{stretch="lin"} is used, then the values are stretched within the range of \code{zlim}. This allows creating consistent coloring between SpatRasters with different cell-value ranges, even when stretching the colors for improved contrast}

\item{stretch}{character. Option to stretch the values to increase contrast: "lin" (linear) or "hist" (histogram). The linear stretch uses \code{\link[terra]{stretch}} with arguments \code{minq=0.02} and \code{maxq=0.98}}

\item{colNA}{color. The color used for cells that have NA values}

\item{label}{a character string to append to the filename}
}
\value{
a list of SpatRaster objects

a list of sources

a character vector of new filenames
}
\description{
print method for ard_glad objects

plot method for ard_glad objects

generic to coerce ard_glad_urls objects to a list

get sources for 'ard_glad' objects

method for default sources - not supported

get sources for 'ard_glad' objects

function to construct new GeoTiff filenames for ard_glad objects

method for default sources - not supported

method for ard_glad sources
}
\details{
This function is used for creating new filenmaes for providing
save locations for functions such as \code{ard_glad_mask}
}
\concept{GLAD ARD utility}
