% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-download.R
\name{ard_glad_download.ard_glad_urls_aws}
\alias{ard_glad_download.ard_glad_urls_aws}
\title{method for AWS sources - not supported}
\usage{
\method{ard_glad_download}{ard_glad_urls_aws}(glad_urls, ...)
}
\description{
method for AWS sources - not supported
}
