% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-download.R
\name{ard_glad_download.ard_glad_urls_umd}
\alias{ard_glad_download.ard_glad_urls_umd}
\title{method for UMD sources}
\usage{
\method{ard_glad_download}{ard_glad_urls_umd}(glad_urls, dir, prefix = dir, return_filename = FALSE)
}
\description{
method for UMD sources
}
