% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-download.R
\name{ard_glad_download.default}
\alias{ard_glad_download.default}
\title{method for default sources - not supported}
\usage{
\method{ard_glad_download}{default}(glad_urls, ...)
}
\description{
method for default sources - not supported
}
