% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-download.R
\name{ard_glad_download}
\alias{ard_glad_download}
\title{Download GLAD ARD data}
\usage{
ard_glad_download(glad_urls, dir, prefix = dir, return_filenmae = FALSE)
}
\arguments{
\item{glad_urls}{a list of URLs to download from \code{ard_glad_urls}}

\item{dir}{a character string of the directory to save the files}

\item{prefix}{a character string to prefix the subdirectory/filenames}

\item{return_filenmae}{a logical value indicating whether to return the
filename or the SpatRaster object}
}
\value{
a list of SpatRaster objects or filenames if \code{return_filenmae} is
TRUE
}
\description{
Download GLAD ARD data
}
