% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-mask.R
\name{ard_glad_mask.ard_glad_src}
\alias{ard_glad_mask.ard_glad_src}
\title{mask method for ard_glad_src objects}
\usage{
\method{ard_glad_mask}{ard_glad_src}(glad, mask_band = "QA", keep_bits = c(1, 2, 15))
}
\description{
mask method for ard_glad_src objects
}
