% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-mask.R
\name{ard_glad_mask.default}
\alias{ard_glad_mask.default}
\title{method for default sources - not supported}
\usage{
\method{ard_glad_mask}{default}(glad, ...)
}
\description{
method for default sources - not supported
}
