% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-urls.R
\name{ard_glad_urls}
\alias{ard_glad_urls}
\title{Generate URLs for GLAD ARD data}
\usage{
ard_glad_urls(
  aoi,
  start_date,
  end_date = NULL,
  rerpoj = TRUE,
  src = c("umd", "aws")
)
}
\arguments{
\item{aoi}{an sf or sfc object representing the area of interest.}

\item{start_date}{a character string representing the start date.}

\item{end_date}{a character string representing the end date. if NULL, only
the start date is used and tiles for a single 16-day interval are returned.}

\item{rerpoj}{a logical value indicating whether to reproject the input if
it is not in the WGS84 CRS.}

\item{src}{a character string indicating the source of the data. Options are
"umd" or "aws".}
}
\value{
a named list of character vectors containing the URLs for the
requested data. names correspond to the start day of the 16-day interval.
}
\description{
Generate URLs for GLAD ARD data
}
\details{
Each item in the returned list comprises the tile(s) for a given 16-day
interval. thelist items are named according to the start date of the
interval.
}
