% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad-tiles.R
\name{gladtiles}
\alias{gladtiles}
\title{function to load the global ARD tiles}
\usage{
gladtiles()
}
\description{
function to load the global ARD tiles
}
\details{
This tile dataset is from
https://glad.umd.edu/users/Potapov/ARD/Global_ARD_tiles.zip
}
\keyword{internal}
