% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-glad-download.R
\name{ard_glad_download}
\alias{ard_glad_download}
\alias{ard_glad_download.default}
\alias{ard_glad_download.ard_glad_urls_aws}
\alias{ard_glad_download.ard_glad_urls_umd}
\title{Download GLAD ARD data}
\usage{
ard_glad_download(glad_urls, dir, prefix = dir, quiet = TRUE, ...)

\method{ard_glad_download}{default}(glad_urls, ...)

\method{ard_glad_download}{ard_glad_urls_aws}(glad_urls, ...)

\method{ard_glad_download}{ard_glad_urls_umd}(glad_urls, dir, prefix = dir, quiet = TRUE, ...)
}
\arguments{
\item{glad_urls}{a list of URLs to download from \code{ard_glad_urls}}

\item{dir}{a character string of the directory to save the files}

\item{prefix}{a character string to prefix the subdirectory/filenames}

\item{quiet}{a logical to suppress messages}

\item{...}{additional arguments not used}
}
\value{
a list of SpatRaster objects or filenames if \code{return_filename} is
TRUE
}
\description{
Download GLAD ARD data

method for default sources - not supported

method for AWS sources - not supported

method for UMD sources
}
\concept{GLAD ARD download}
